wip with Maury on Options
