+++ src/options.tsx
@@ -14,69 +14,23 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import { IconDefinition } from '@fortawesome/fontawesome-svg-core'
-import { faArrowLeft, faArrowRight, faCog, faPlay, faQuestionCircle } from '@fortawesome/free-solid-svg-icons'
-import {
-    NxGlobalSidebarFooter,
-    NxGlobalSidebarNavigation,
-    NxGlobalSidebarNavigationLink,
-    NxPageMain,
-    NxStatefulGlobalSidebar,
-} from '@sonatype/react-shared-components'
+
 import React from 'react'
 import ReactDOM from 'react-dom/client'
-import { NexusOptionsContainer } from './NexusOptionsContainer'
 import { logger, LogLevel } from './logger/Logger'
 import { MESSAGE_REQUEST_TYPE } from './types/Message'
 import { readExtensionConfiguration } from './messages/SettingsMessages'
+import Options from './components/Options/Options'
 
 // eslint-disable-next-line @typescript-eslint/strict-boolean-expressions, @typescript-eslint/no-explicit-any
 const _browser: any = chrome ? chrome : browser
-const extension = _browser.runtime.getManifest()
 
 const container = document.getElementById('ui')
 const root = ReactDOM.createRoot(container)
-const search = window.location.search
-const params = new URLSearchParams(search)
-const install = params.has('install')
-const help = params.has('help')
 
 root.render(
     <React.StrictMode>
-        <NxStatefulGlobalSidebar
-            isDefaultOpen={true}
-            toggleCloseIcon={faArrowRight as IconDefinition}
-            toggleOpenIcon={faArrowLeft as IconDefinition}
-            logoImg='images/sonatype-lifecycle-logo-nav-white.svg'
-            logoAltText={_browser.i18n.getMessage('EXTENSION_NAME')}
-            logoLink='#'>
-            <NxGlobalSidebarNavigation>
-                <NxGlobalSidebarNavigationLink
-                    icon={faPlay as IconDefinition}
-                    text={_browser.i18n.getMessage('SIDEBAR_LINK_GETTING_STARTED')}
-                    href='options.html?install'
-                />
-                <NxGlobalSidebarNavigationLink
-                    icon={faQuestionCircle as IconDefinition}
-                    text={_browser.i18n.getMessage('SIDEBAR_LINK_HELP')}
-                    href='options.html?help'
-                />
-                <NxGlobalSidebarNavigationLink
-                    icon={faCog as IconDefinition}
-                    text={_browser.i18n.getMessage('SIDEBAR_LINK_OPTIONS')}
-                    href='options.html'
-                />
-            </NxGlobalSidebarNavigation>
-            <NxGlobalSidebarFooter
-                supportText={_browser.i18n.getMessage('SIDEBAR_FOOTER_LINK_REQUEST_SUPPORT')}
-                supportLink={extension.homepage_url}
-                releaseText={_browser.i18n.getMessage('RELEASE_VERSION', extension.version)}
-                showCreatedBy={true}
-            />
-        </NxStatefulGlobalSidebar>
-        <NxPageMain>
-            <NexusOptionsContainer install={install} help={help} />
-        </NxPageMain>
+        <Options />
     </React.StrictMode>
 )
